requirements.txt
snowflake-snowpark-python[pandas]==1.5.1
snowflake-connector-python==2.7.9
streamlit==1.24.0
openai==0.27.8
langchain==0.0.305
pandas==1.5.0
pydantic==1.10.8
supabase==1.0.3
unstructured==0.7.12
tiktoken==0.4.0
black==23.3.0
replicate==0.8.4
boto3==1.28.57
snowflake-sqlalchemy-1.5.0
typing-extensions>=4.3.0